fix: Tools export | historical tokens docs
