+++ packages/ai-agent-sdk/src/core/agent/index.ts
@@ -2,7 +2,7 @@ import { assistant, Base, system, user } from "../base";
 import type { ModelConfig } from "../llm";
 import { LLM } from "../llm";
 import { StateFn, type ZeeWorkflowState } from "../state";
-import type { Tool } from "../tools";
+import type { Tool } from "../tools/base";
 import type { ParsedFunctionToolCall } from "openai/resources/beta/chat/completions";
 import type { ChatCompletionMessageParam } from "openai/resources/chat/completions";
 import z from "zod";
