+++ packages/ai-agent-sdk/src/core/tools/goldrush/index.ts
@@ -1,3 +1,5 @@
 export * from "./token-balances";
 export * from "./nft-balances";
 export * from "./transactions";
+export * from "./historical-token-price";
+export * from "./base";
