+++ packages/ai-agent-sdk/src/core/zee/index.test.ts
@@ -1,5 +1,5 @@
 import { Agent } from "../agent";
-import { createTool } from "../tools";
+import { createTool } from "../tools/base";
 import { ZeeWorkflow } from "./index";
 import fetch from "node-fetch";
 import { test } from "vitest";
